rework method of picture comparison
I rethought and rewrote the method of comparing a picture with the target
picture, which I was previously doing  with histograms. Really, all I'm
trying to do is determine the raw difference between one matrix and
another, so I'm just subtracting the evo_pic matrix from the target matrix
and summing the resultant matrix. In theory, the more different an
evolutionary picture is from the target, the greater the resulting
difference.

Makes sense, I think.

Also, reworked display_pic() to be able to accept and display a list of
picture-like matrices.