add evolution iteration method
Addedd  the Evolver.iterate_evo() method, which runs one generation of the
GA. It runs hill-climber style, selecting the top few Pictures in each
generation and creating a new population by mutating them. Each surviving
Picture is bred with each other surviving picture (including itself) and
is mutated by itself a number of times equal to the number of surviving
Pictures.

Also, I changed the break-down of populations, so that the total
population is x + 2x^2, where x is the number of surviving Pictures. The
value of x is held in a constant in the Evolver class.