add command-line arguments
First argument is how many generations to run, the second is how many
generations towait before saving pictures. A 0 in the first argument
indicates running until the survivors are below a certain fitness
threshold.

I'll probably change this later so that the algorithm always runs until
its below a certain threshold and the first argument dictates the
threshold.